+++ src/ServiceProvider.php
@@ -5,6 +5,7 @@ namespace Surgems\RedirectUrls;
 use Statamic\Providers\AddonServiceProvider;
 use Surgems\RedirectUrls\Middleware\Handle404;
 use Statamic\Statamic;
+use Statamic\Facades\CP\Nav;
 
 class ServiceProvider extends AddonServiceProvider
 {
@@ -14,26 +15,42 @@ class ServiceProvider extends AddonServiceProvider
         ],
     ];
 
+    protected $routes = [
+        'cp' => __DIR__.'/../routes/cp.php',
+    ];
+
     public function bootAddon()
     {
         $this->registerAddonConfig();
-
-        Statamic::afterInstalled(function ($command) {
-            $file_path = 'config/redirect-urls.php';
-            $template = include(__DIR__.'/templates/redirect-urls-array.php');
-
-            if(! file_exists($file_path))
-            {
-                file_put_contents($file_path, $template);
-            }
-
-            return $this;
-        });
+        $this->bootAddonViews();
+        $this->bootAddonNav();
     }
 
     protected function registerAddonConfig()
     {
         $this->mergeConfigFrom(__DIR__.'/../config/config.php', 'statamic.redirect-urls');
+
+        return $this;
+    }
+
+    protected function bootAddonViews()
+    {
+        $this->loadViewsFrom(__DIR__.'/../resources/views', 'redirect-urls');
+
+        return $this;
+    }
+
+
+    protected function bootAddonNav()
+    {
+        Nav::extend(function ($nav) {
+            $nav->tools('Redirect Urls')
+                ->route('redirect-urls.dashboard')
+                ->icon('git')
+                ->children([
+                    'Import Redirects' => cp_route('redirect-urls.import'),
+                ]);
+        });
 
         return $this;
     }