+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [Release] - 2022-05-26
 
 ``` bash
-composer require surgems/redirect-url:1.0
+composer require surgems/redirect-urls:1.0
 ```
  
 ### Added
@@ -18,7 +18,7 @@ composer require surgems/redirect-url:1.0
 ## [1.1] - 2022-05-27
 
 ``` bash
-composer require surgems/redirect-url:1.1
+composer require surgems/redirect-urls:1.1
 ```
  
 ### Added
@@ -39,7 +39,7 @@ composer require surgems/redirect-url:1.1
 ## [1.1.1] - 2022-05-27
 
 ``` bash
-composer require surgems/redirect-url:1.1.1
+composer require surgems/redirect-urls:1.1.1
 ```
  
 ### Added
@@ -50,3 +50,22 @@ composer require surgems/redirect-url:1.1.1
 
 - [ServiceProvider.php](https://github.com/JacobTinston/StatamicRedirectUrls/blob/master/src/ServiceProvider.php)
   Use __statamic.web__ middleware group insted of __web__ as it was interfering with image requests.
+
+
+## [1.2] - 2022-05-29
+
+``` bash
+composer require surgems/redirect-urls
+```
+ 
+### Added
+
+- CSV Importer
+- Redirect Urls CP Nav
+ 
+### Changed
+
+- Removed the hardcoded array
+- Use YAML instead of PHP arrays to store redirects
+ 
+### Fixed